Moves duplicate route utils to the same file
